t-ex_core.as
actually i lost the source once and decompiled it myself and went on doing some small modifications on it... so the local vars have lost their names. but i'm mostly using global variables, so..... not so bad (slap
note the current block has to be "failing to perform requested move" ("pushed on the wall") to use wktable[1] or [2], not only "the move button is pushed".